% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissmodelr-package.R
\docType{package}
\name{dissmodelr-package}
\alias{dissmodelr}
\alias{dissmodelr-package}
\title{dissmodelr: Succulent simulation model}
\description{
The package contains the succulent simulation model from Reineking et al. 2006, and code to run simulations for the simnet project.
}
\author{
\strong{Maintainer}: Björn Reineking \email{bjoern.reineking@inrae.fr} (\href{https://orcid.org/0000-0001-5277-9181}{ORCID})

}
\keyword{internal}
