% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_configuration.R
\name{make_configuration}
\alias{make_configuration}
\title{Create succulent model configuration text file}
\usage{
make_configuration(
  repetitions = 1,
  gridWidth = 20,
  gridHeight = 20,
  cellwidth = 0.25,
  numyears = 500,
  timeStepsPerYear = 365,
  timestepLength = 1,
  soilDepth = 0.3,
  thetaSat = 0.4,
  kappa = 0.7,
  mu = -1,
  sd = 0.8,
  seedrainIntensity = 30,
  seedrainYears = 100,
  seedMortality = 0.3,
  maxIndividualDensity = 499.9,
  sensitivityAnalysis = "false",
  initSoilWaterContent = 0.2,
 
    redistributionFactorFile = system.file("input/diversityMechanisms5/heterogeneousLandscapeSmall.txt",
    package = "dissmodelr"),
  dailyRainFile = system.file("input/diversityMechanisms5/rain90mm5000.dat", package =
    "dissmodelr"),
  plantInitFile = "none",
  plantParsFile = system.file("input/diversityMechanisms5/PlantPars.dat", package =
    "dissmodelr"),
  metacommunityFile = "none",
  outputFilename = "outSim17",
  outputOptions = "popsizeseedscoverbiomassswcdeathwateruptake",
  reportCycle = 1825,
  dailyEvapoFile = system.file("input/diversityMechanisms5/numeesSouthEvapo.dat",
    package = "dissmodelr"),
  simulationType = "complex",
  initSeedbankDensity = 3000,
  rocksAndPotsFile = system.file("input/diversityMechanisms5/pots20small.txt", package =
    "dissmodelr"),
  thetaCrit = 0.04,
  dewAmount = 0.1,
  dewDays = 200
)
}
\arguments{
\item{repetitions}{Number of repetitions to do. UNUSED}

\item{gridWidth}{Number of grid columns}

\item{gridHeight}{Number of grid rows}

\item{cellwidth}{Length of cells (cm)}

\item{numyears}{Number of years to simulate}

\item{timeStepsPerYear}{Number of time steps per year to simulate}

\item{timestepLength}{Duration of individual time step (days)}

\item{soilDepth}{Depth of soil layer (m)}

\item{thetaSat}{Soil water saturation level (fraction)}

\item{kappa}{TODO}

\item{mu}{TODO}

\item{sd}{TODO}

\item{seedrainIntensity}{Number of external seeds (sum across all species) (m-2)}

\item{seedrainYears}{Number of years during with external seedrain (starts in first year)}

\item{seedMortality}{TODO}

\item{maxIndividualDensity}{TODO}

\item{sensitivityAnalysis}{TODO}

\item{initSoilWaterContent}{TODO}

\item{redistributionFactorFile}{TODO}

\item{dailyRainFile}{TODO}

\item{plantInitFile}{TODO}

\item{plantParsFile}{TODO}

\item{metacommunityFile}{TODO}

\item{outputFilename}{TODO}

\item{outputOptions}{TODO}

\item{reportCycle}{TODO}

\item{dailyEvapoFile}{TODO}

\item{simulationType}{TODO}

\item{initSeedbankDensity}{TODO}

\item{rocksAndPotsFile}{TODO}

\item{thetaCrit}{TODO}

\item{dewAmount}{TODO}

\item{dewDays}{TODO}
}
\value{
Text with configuration file
}
\description{
Create succulent model configuration text file
}
