% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simnet_simulation_df.R
\name{make_simnet_simulation_df}
\alias{make_simnet_simulation_df}
\title{Make simulation data frame for simnet}
\usage{
make_simnet_simulation_df(
  Model = "bjoern",
  Ninitial = as.integer(2^(0:6)),
  Rep = 1:64,
  SeedRain = c(0, 5, 10, 50, 100, 1000),
  Stage = c("assembly"),
  Wdh = 5
)
}
\arguments{
\item{Model}{Name of model}

\item{Ninitial}{Initial number of species}

\item{Rep}{Replicate of species pool}

\item{SeedRain}{Percentage of reference seed rain}

\item{Stage}{Simulation stage}

\item{Wdh}{Stochastic replicate for given species pool}
}
\value{
Data frame with simulation parameter values
}
\description{
Make simulation data frame for simnet
}
