% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_pools.R
\name{make_species_pools}
\alias{make_species_pools}
\title{Make simulation pools}
\usage{
make_species_pools(simulations)
}
\arguments{
\item{simulations}{Data frame with simulation parameters}
}
\value{
Named list with species pools
}
\description{
Make simulation pools
}
