% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plantpars.R
\name{read_plantpars}
\alias{read_plantpars}
\title{Read succulent model plant parameter file}
\usage{
read_plantpars(file)
}
\arguments{
\item{file}{File containing plant parameter values}
}
\value{
Data frame with plant parameter values
}
\description{
Read succulent model plant parameter file
}
