% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simnet.R
\name{run_simnet}
\alias{run_simnet}
\title{Run simnet experiment}
\usage{
run_simnet(
  wdh,
  base_path = "~/simnet",
  mean_annual_seedrain_monoculture_m2 = 19402,
  test = -1,
  mcseed = 3
)
}
\arguments{
\item{wdh}{Number of replicates}

\item{base_path}{Base path for writing files}

\item{mean_annual_seedrain_monoculture_m2}{Number of seeds}

\item{test}{Number of simulations to test (negative values mean use all)}

\item{mcseed}{Random number seed for simulation}
}
\value{
Data frame with plant parameter values
}
\description{
Run simnet experiment
}
