% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate}
\alias{simulate}
\title{Runs dissmodel}
\arguments{
\item{expParFilename}{Name of input file}
}
\value{
Time series of biomass and productivity of all species
}
\description{
Runs dissmodel
}
