% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plantpars.R
\name{write_plantpars}
\alias{write_plantpars}
\title{Write succulent model plant parameter file}
\usage{
write_plantpars(x, file)
}
\arguments{
\item{x}{Data framw with plant parameters}

\item{file}{Name of file to create}
}
\description{
Write succulent model plant parameter file
}
